Knowledge Graphs for RAG/requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple/

langchain_community>=0.0.10
langchain_openai>=0.1.0
python-dotenv>=1.0.0
neo4j>=5.0.0
langchain>=0.1.0
openai>=1.0.0 